DELETE: add recursive option

+++ www/wildcard/DELETE.php
@@ -5,6 +5,22 @@
  * $Id$
  */
 
+function rrmdir($dir) {
+    if (is_dir($dir)) {
+        $objects = scandir($dir);
+        foreach ($objects as $object) {
+            if ($object != "." && $object != "..") {
+                if (filetype($dir."/".$object) == "dir")
+                    rrmdir($dir."/".$object);
+                else
+                    unlink($dir."/".$object);
+            }
+        }
+        reset($objects);
+        rmdir($dir);
+    }
+}
+
 require_once('runtime.php');
 
 // permissions
@@ -25,7 +41,10 @@ if ($frag) {
 }
 
 if (is_dir($_filename)) {
-    rmdir($_filename);
+    if ($_options->recursive)
+        rrmdir($_filename);
+    else
+        rmdir($_filename);
 } elseif (file_exists($_filename)) {
     unlink($_filename);
 } else {
